changed maxRecords counter in B2C from int to long

+++ src/main/java/net/sf/cram/Bam2Cram.java
@@ -438,7 +438,7 @@ public class Bam2Cram {
 		File outputCramFile = null;
 
 		@Parameter(names = { "--max-records" }, description = "Stop after compressing this many records. ")
-		int maxRecords = Integer.MAX_VALUE;
+		long maxRecords = Long.MAX_VALUE;
 
 		@Parameter
 		List<String> sequences;
